refactor: sorted mode and modules into their own packages and cleaned up code
