+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/movement/ModuleNoWeb.kt
@@ -33,17 +33,14 @@ import net.ccbluex.liquidbounce.utils.client.notification
  */
 object ModuleNoWeb : Module("NoWeb", Category.MOVEMENT) {
 
-    val modes = choices("Mode", { Air }) {
-        arrayOf(
-            Air
-        )
-    }
+    val modes = choices("Mode", Air, arrayOf(Air))
 
     val repeatable = repeatable {
         if (ModuleAvoidHazards.enabled) {
             if (ModuleAvoidHazards.cobWebs) {
                 ModuleAvoidHazards.enabled = false
-                notification("Compatibility error", "NoWeb is incompatible with AvoidHazards", NotificationEvent.Severity.ERROR)
+                notification("Compatibility error", "NoWeb is incompatible with AvoidHazards",
+                    NotificationEvent.Severity.ERROR)
             }
         }
     }
