+++ src/main/resources/assets/liquidbounce/lang/en_pt.json
@@ -163,7 +163,7 @@
   "liquidbounce.module.parkour.description": "Autwomaticawwy jumps at teh wewwy edge of a bwock",
   "liquidbounce.module.speed.description": "Makes u gwo fast",
   "liquidbounce.module.strafe.description": "Awwows u two stwafe intwo diffewent diwections whiwal u awe in mid-aiw",
-  "liquidbounce.module.vehicleFly.description": "Awwows u two fwy fwough teh aiw but using a wehicwal ｡･:*:･ﾟ★,｡･:*:･ﾟ☆mwountabwal entity☆ﾟ･:*:･｡,★ﾟ･:*:･｡",
+  "liquidbounce.module.vehicleControl.description": "Awwows u two fwy fwough teh aiw but using a wehicwal ｡･:*:･ﾟ★,｡･:*:･ﾟ☆mwountabwal entity☆ﾟ･:*:･｡,★ﾟ･:*:･｡",
   "liquidbounce.module.antiCactus.description": "Pwewents taking damage fwom cacti",
   "liquidbounce.module.antiLevitation.description": "Stwops teh wevitation and swow fawwing effect",
   "liquidbounce.module.autoBreak.description": "Autwomaticawwy bweaks bwocks",
